Created new component for rendering data in signature-request-message.js
