+++ ui/components/app/signature-request/index.scss
@@ -1,6 +1,7 @@
 @import 'signature-request-footer/index';
 @import 'signature-request-header/index';
 @import 'signature-request-message/index';
+@import 'signature-request-data/index';
 
 .signature-request {
   display: flex;
