+++ ui/components/app/signature-request/signature-request-data/index.js
@@ -1,0 +1,1 @@
+export { default } from './signature-request-data';
