+++ ui/components/app/signature-request/signature-request-message/index.scss
@@ -12,31 +12,6 @@
     }
   }
 
-  &__node {
-    &__value {
-      white-space: pre-line;
-      overflow: hidden;
-      word-wrap: break-word;
-
-      &__address {
-        [dir='rtl'] & {
-          /*rtl:ignore*/
-          direction: ltr;
-
-          /*rtl:ignore*/
-          text-align: right;
-
-          span {
-            display: block;
-
-            /*rtl:ignore*/
-            direction: rtl;
-          }
-        }
-      }
-    }
-  }
-
   &__scroll-button {
     position: absolute;
     right: 28px;
